Fixed bug in GUI 2 (incorrect handling for special case)
There are always 14 periods for 7-8 days returned by the API, so the previous handling was incorrect.
Instead, if the first period is "Tonight" (meaning there is only one entry for the current day), we handle it manually and shift the period indices for the rest of the code to compensate.