Fixed bug which offset forecast days by 1
The handling from the last commit didn't work either, because sometimes the first period is "Tonight", other times it's "Overnight", etc...
I'm guessing a surefire way to tell if we're missing an entry is if the second period is a full day (e.g. Monday, Tuesday, as opposed to Monday Night, Tuesday Night). If we're missing an entry, apply our previous handling as normal.